HktGameplayGenerator/McpServer/pyproject.toml
[project]
name = "hkt-mcp"
version = "0.1.0"
description = "MCP Server for Unreal Engine 5 integration with LLM clients"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "HKT"}
]
dependencies = [
    "mcp>=1.0.0",
    "websockets>=12.0",
    "pydantic>=2.0.0",
    "aiohttp>=3.9.0",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
hkt-mcp = "hkt_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/hkt_mcp"]

[tool.pytest.ini_options]
asyncio_mode = "auto"